refactor: set up folder structure (#17)
