+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+
+import logo from './assets/logo.svg';
+import './styles/App.css';
 
 const App = () => {
   return (
